chore(guide) remove automatic recentering map when location is received
