+++ app/src/main/res/drawable/bg_white_rounded_top.xml
@@ -8,8 +8,8 @@
         <shape xmlns:android="http://schemas.android.com/apk/res/android"
                android:shape="rectangle">
             <corners
-                android:topLeftRadius="10dp"
-                android:topRightRadius="10dp"
+                android:topLeftRadius="15dp"
+                android:topRightRadius="15dp"
                 android:bottomLeftRadius="0dp"
                 android:bottomRightRadius="0dp"/>
             <stroke android:color="@color/light_grey" android:width="1dp"/>
